// frontend/src/api/inventoryService/model/inventoryItemResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface InventoryItemResponse {
  id?: number;
  productId?: string;
  productName?: string;
  quantity?: number;
  reservedQuantity?: number;
  availableQuantity?: number;
  unit?: string;
}